- reduced the size of a header flit, from 22 bytes, to 10 bytes; - bug fix: flit header was removed from data flits too (this problem was visible only when I decreased the header size); - checking a flit if it's a header flit is more reliable if the NoC packet tag is used.

+++ src/noc/packet/noc-packet.cc
@@ -29,8 +29,8 @@ NS_LOG_COMPONENT_DEFINE ("NocPacket");
 namespace ns3
 {
 
-  NocPacket::NocPacket (uint32_t xDistance, uint32_t yDistance, uint32_t sourceX,
-      uint32_t sourceY, uint16_t dataFlitCount, uint32_t dataPacketSize) :
+  NocPacket::NocPacket (uint8_t xDistance, uint8_t yDistance, uint8_t sourceX,
+      uint8_t sourceY, uint16_t dataFlitCount, uint32_t dataPacketSize) :
     Packet (dataPacketSize)
   {
     NocHeader nocHeader (xDistance, yDistance, sourceX, sourceY, dataFlitCount);
